Add Basic files info for instructions Add Basic Setup for Address and Info

+++ src/lib/vm.rs
@@ -1,7 +1,7 @@
 use crate::arch::i8086;
 use std::boxed::Box;
 
-pub const MB: usize = 1 * 1024 * 1024;
+pub const MB: u32 = 1 * 1024 * 1024;
 
 // All these are arbitrary
 pub const DATA_SEG_BEGIN: u16 = 0x7000; // 28672
@@ -10,14 +10,14 @@ pub const STACK_SEG_BEGIN: u16 = 0x1000; // 4096
 // TODO what to use for mem
 pub struct VM {
     pub arch: i8086,
-    pub mem: Box<[u8; MB]>,
+    pub mem: Box<[u8; MB as usize]>,
 }
 
 impl VM {
     pub fn new() -> VM {
         let mut ret = VM {
             arch: i8086::default(),
-            mem: Box::new([0; MB]),
+            mem: Box::new([0; MB as usize]),
         };
 
         // Set default values for registers
